Some bug fixes making for a smooth game-run
Fixed a glitch with dividebyzero in Cards.cpp, and made some changes to the output from order.cpp to enhance readability in the console. Made some modification in GameEngine to make every alternate player aggressive and benevolent. Made (temporarily) PlayerStrategy and a player's OrderList public(OrderList is public as the list is easier to clear the list after every turn directly on the Game Engine) and lastly after we issue an attack order with AggressivePlayer, we set the temp armycount to 0 so that the issueOrder() knows the turn is over.